requirements.txt
discord.py[voice]
flask